feat(scripts): ✨ add TRACE constraint to solve_network
